+++ packages/ap2/src/index.js
@@ -1,8 +1,18 @@
 export * from './receipts';
-export { computeSdHash } from '@docknetwork/crypto-utils/vc';
+export * from './mandates';
+export * from './signer';
+export { computeSdHash, parseSdJwtPresentation } from '@docknetwork/crypto-utils/vc';
+export { Secp256r1Keypair } from '@docknetwork/crypto-utils/keypairs';
 export {
   encodeDisclosure,
   inferReceiptType,
   validateReceipt,
   validateReceiptTime,
+  validateMandateContent,
+  checkoutReceiptSchema,
+  paymentReceiptSchema,
+  checkoutMandateOpenSchema,
+  checkoutMandateClosedSchema,
+  paymentMandateOpenSchema,
+  paymentMandateClosedSchema,
 } from './utils';